Refactoring and conversion to .hpp files
