+++ Firmware code/.editorconfig
