+++ Firmware code/include/global_defines.hpp
