+++ Firmware code/include/smart_energy_meter.h
@@ -1,3 +1,4 @@
+#pragma once
 // defines
 #define TINY_GSM_MODEM_SIM800
 #define TINY_GSM_RX_BUFFER 60  // Set RX buffer to 1Kb
@@ -21,8 +22,7 @@
 
 // OpenSmartMeter libraries
 #include "SAM_UART.h"
-#include "global_defines.hpp"
-#include "relay.hpp"
+#include "global_defines.h"
 
 HardwareSerial Serial2(PA3, PA2);
 HardwareSerial ATM90E26(PB11, PB10);
@@ -31,7 +31,8 @@ ATM90E26_UART AFE_chip(&ATM90E26);
 // keypad
 #define Lengths 20
 char Data[Lengths];
-int lcd_count, convertedsts_data = 0;
+int lcd_count = 0;
+unsigned long convertedsts_data = 0.0;
 byte data_count, data_count2, dt, sts_accept, encoder = 0;
 char customKey;
 const byte ROWS = 4;
@@ -124,6 +125,7 @@ unsigned long token_eeprom_location = 20;
 unsigned long eeprom_location_cnt = 40;
 unsigned long credit_eeprom_location = 9;
 
+
 //    for meter
 String meter = "MT100";
 unsigned long meter_no = 100;
@@ -143,127 +145,3 @@ byte fe2[8] = {0b11111, 0b11111, 0b00000, 0b00000,
 byte fe4[8] = {0b01111, 0b01100, 0b01100, 0b00000,
                0b00000, 0b00000, 0b00000, 0b00000};
 
-void setup() {
-  lcd.begin(16, 2);
-  pinMode(buzzer, OUTPUT);
-  pinMode(relaya, OUTPUT);
-  pinMode(relayb, OUTPUT);
-  pinMode(red_led, OUTPUT);
-  pinMode(green_led, OUTPUT);
-
-  digitalWrite(buzzer, HIGH);
-
-  lcd.createChar(0, fe1);
-  lcd.createChar(1, fe2);
-  lcd.createChar(2, fe3);
-  lcd.createChar(3, fe4);
-  lcd.setCursor(0, 0);
-  lcd.print("  DEVELOPED BY  ");
-  lcd.setCursor(0, 1);
-  lcd.print(" FIRST ELECTRIC ");
-
-  lcd.clear();
-  lcd.setCursor(7, 0);
-  lcd.write(byte(0));
-  lcd.setCursor(8, 0);
-  lcd.write(byte(1));
-  lcd.setCursor(7, 1);
-  lcd.write(byte(2));
-  lcd.setCursor(8, 1);
-  lcd.write(byte(3));
-
-  ATM90E26.begin(9600);
-  AFE_chip.SET_register_values();
-  delay(1000);
-  digitalWrite(buzzer, LOW);
-
-  Serial2.begin(115200);
-  Serial2.print("working");
-  Serial1.begin(115200);
-  delay(3000);
-  Serial1.write("AT+IPR=9600\r\n");
-  Serial1.end();
-  Serial1.begin(9600);
-
-  lcd.setCursor(0, 0);
-  lcd.print(" SYSTEM BOOTING ");
-  lcd.setCursor(0, 1);
-  lcd.print(" ##             ");
-
-  Serial2.println("Initializing modem...");
-  modem.restart();
-  String modemInfo = modem.getModemInfo();
-  lcd.setCursor(0, 0);
-  lcd.print("                     ");
-  lcd.setCursor(0, 1);
-  lcd.print("                    ");
-
-  Serial2.print("Modem: ");
-  Serial2.println(modemInfo);
-  lcd.setCursor(0, 0);
-  lcd.print(" modemInfo ");
-  lcd.setCursor(0, 1);
-  lcd.print(modemInfo);
-  delay(2000);
-  lcd.clear();
-
-  if (!rtc.begin()) {
-    lcd.print("Couldn't find RTC");
-    delay(2000);
-    while (1)
-      ;
-  }
-  if (!rtc.isrunning()) {
-    lcd.print("RTC is NOT running!");
-    delay(2000);
-  }
-
-  lcd.setCursor(0, 0);
-  lcd.print("CSOne :         ");
-  lcd.setCursor(8, 0);
-  lcd.print(AFE_chip.FETCH_MeterCSOne());
-
-  lcd.setCursor(0, 1);
-  lcd.print("CSTwo :         ");
-  lcd.setCursor(8, 1);
-  lcd.print(AFE_chip.FETCH_MeterCSTwo());
-  delay(1000);
-
-  // rtc eeprom
-  // mem.writeLong(credit_eeprom_location, 200);
-  //  mem.writeLong(eeprom_location_cnt, token_eeprom_location);
-  //  delay(20);
-  //  mem.writeLong(token_eeprom_location, 200);
-  //  delay(20);
-  creditt = mem.readLong(credit_eeprom_location);
-  if (creditt >= 0) {
-    mem.writeLong(credit_eeprom_location, creditt);
-  }
-  delay(10);
-  relay_on();
-
-#if defined(TIM1)
-  TIM_TypeDef* Instance = TIM1;
-#else
-  TIM_TypeDef* Instance = TIM2;
-#endif
-  HardwareTimer* MyTim = new HardwareTimer(Instance);
-  MyTim->setOverflow(20, HERTZ_FORMAT);
-  MyTim->attachInterrupt(urgeent);
-  MyTim->resume();
-}
-
-void loop() {
-  mesure();
-  if ((mains_input_value > 50)) {
-    credit_reminder();
-  }
-  if ((mains_input_value < 50)) {
-    digitalWrite(red_led, LOW);
-    digitalWrite(green_led, LOW);
-  }
-  get_time();
-  if ((sts_mode == 0) && (mains_input_value > 50)) {
-    gsm_func();
-  }
-}
