+++ Firmware code/include/sts_token.hpp
@@ -1,3 +1,6 @@
+#pragma once
+#include "smart_energy_meter.h"
+
 void STStoken_decode() {
   String keycofirmation = sts_data.substring(0, 3);
   int convertekeycofirmation = keycofirmation.toInt();
@@ -7,7 +10,7 @@ void STStoken_decode() {
 
   String sts_meter_no_count = sts_data.substring(4, 5);
   convertedsts_data = sts_data.toInt();
-  int stsnew_meter_no_count = sts_meter_no_count.toInt();
+  unsigned int stsnew_meter_no_count = sts_meter_no_count.toInt();
   // int stsnew_meter_no_count = sts_meter_no_count;
   String sts_day = sts_data.substring(6, 8);
   convertedsts_day = sts_day.toInt();
@@ -17,8 +20,8 @@ void STStoken_decode() {
   String identifier = sts_third.substring(6, 7);
   String sts_encode = sts_data.substring(15, 20);
   long tariff_MT_NO = sts_encode.toInt();
-  long tariff_gotten = ((tariff_MT_NO) - (meter_no * multiplier));
-  long trueMT_NO = tariff_MT_NO - tariff;
+  // long tariff_gotten = ((tariff_MT_NO) - (meter_no * multiplier));
+  unsigned long trueMT_NO = tariff_MT_NO - tariff;
   trueMT_NO = trueMT_NO / multiplier;
 
   if ((confirmkey == private_stskey) &&
@@ -153,3 +156,4 @@ void STStoken_decode() {
     digitalWrite(buzzer, HIGH);
   }
 }
+
